# docs/api/openapi.m3.yaml
openapi: 3.1.0
info:
  title: Zalo Shop M3 Inventory, Search, Pricing and Cart API
  version: 0.3.4-m3.6
  description: >-
    M3.1 冻结契约；作为 openapi.m1.yaml 与 openapi.m2.yaml 的向后兼容增量。
    商城只从受信 X-Store-Code 与认证主体解析，请求体不接受 store_id。
    M3.2-M3.5 已实施数据库基础、库存、搜索、促销、优惠券、可信报价与管理员预览；M3.6 已实施会员购物车与移动端接入。
    本契约不进入 M4 订单、支付、物流或售后。
servers:
  - url: http://localhost:3000
...
paths:
...
  /v1/admin/promotions:
    get:
      tags: [Admin promotions]
      summary: 查询当前商城促销
      description: 需要 `store.promotions.read`。
      security: [{ StoreCode: [], AdminBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/AdminLimit'
        - { name: status, in: query, schema: { $ref: '#/components/schemas/PromotionStatus' } }
      responses:
        '200':
          description: 促销分页
          content: { application/json: { schema: { $ref: '#/components/schemas/PromotionPage' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Admin promotions]
      summary: 创建当前商城促销根
      description: 需要 `store.promotions.manage`。
      security: [{ StoreCode: [], AdminBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties: { code: { $ref: '#/components/schemas/Code' } }
      responses:
        '201':
          description: 已创建草稿根
          content: { application/json: { schema: { $ref: '#/components/schemas/Promotion' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/admin/promotions/targets:
    get:
      tags: [Admin promotions]
      summary: 查询当前商城促销目标候选
      description: >-
        需要 `store.promotions.read`。仅返回当前商城的品牌、类目、商品或 SKU
        的编码与三语名称；不接受 STORE 目标，也不返回其他商城资源。
      security: [{ StoreCode: [], AdminBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/AdminLimit'
        - {
            name: target_type,
            in: query,
            required: true,
            schema: { type: string, enum: [BRAND, CATEGORY, PRODUCT, SKU] },
          }
        - { name: q, in: query, schema: { type: string, minLength: 1, maxLength: 100 } }
      responses:
        '200':
          description: 同商城目标候选分页
          content:
            {
              application/json:
                { schema: { $ref: '#/components/schemas/PromotionTargetLookupPage' } },
            }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/admin/promotions/{promotionId}/versions:
    get:
      tags: [Admin promotions]
      summary: 查询当前商城促销的草稿与发布版本
      description: 需要 `store.promotions.read`；发布版本只读。
      security: [{ StoreCode: [], AdminBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/PromotionId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/AdminLimit'
      responses:
        '200':
          description: 版本分页
          content:
            { application/json: { schema: { $ref: '#/components/schemas/PromotionVersionPage' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Admin promotions]
      summary: 创建或替换下一草稿版本
      description: 需要 `store.promotions.manage`；目标必须全部属于当前商城。
      security: [{ StoreCode: [], AdminBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/PromotionId'
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/PromotionVersionInput' } } }
      responses:
        '201':
          description: 草稿版本
          content:
            { application/json: { schema: { $ref: '#/components/schemas/PromotionVersion' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/admin/promotions/{promotionId}/publish:
    post:
      tags: [Admin promotions]
      summary: 发布不可变促销版本
      description: >-
        需要 `store.promotions.publish`、二次确认、乐观锁与 Idempotency-Key。
        发布当前 active 的已发布版本可将同一促销从 PAUSED 恢复为 ACTIVE；不得借此恢复 ENDED
        或切回其他历史发布版本。
      security: [{ StoreCode: [], AdminBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/PromotionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/PublishPromotionRequest' } } }
      responses:
        '200':
          description: 已发布或幂等重放
          content: { application/json: { schema: { $ref: '#/components/schemas/Promotion' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/admin/promotions/{promotionId}/pause:
    post:
      tags: [Admin promotions]
      summary: 暂停当前促销
      description: 需要 `store.promotions.publish`、二次确认、乐观锁与 Idempotency-Key。
      security: [{ StoreCode: [], AdminBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/PromotionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/PausePromotionRequest' } } }
      responses:
        '200':
          {
            description: 已暂停,
            content: { application/json: { schema: { $ref: '#/components/schemas/Promotion' } } },
          }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/admin/promotions/{promotionId}/end:
    post:
      tags: [Admin promotions]
      summary: 不可逆结束当前促销
      description: 需要 `store.promotions.publish`、二次确认、乐观锁与 Idempotency-Key。
      security: [{ StoreCode: [], AdminBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/PromotionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/EndPromotionRequest' } } }
      responses:
        '200':
          {
            description: 已结束,
            content: { application/json: { schema: { $ref: '#/components/schemas/Promotion' } } },
          }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/admin/coupons:
    get:
      tags: [Admin promotions]
      summary: 查询当前商城优惠券
      description: 需要 `store.promotions.read`。
      security: [{ StoreCode: [], AdminBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/AdminLimit'
      responses:
        '200':
          description: 优惠券分页
          content: { application/json: { schema: { $ref: '#/components/schemas/CouponPage' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Admin promotions]
      summary: 创建当前商城优惠券草稿
      description: 需要 `store.promotions.manage`；引用已发布 COUPON 版本。
      security: [{ StoreCode: [], AdminBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CouponInput' } } }
      responses:
        '201':
          description: 已创建
          content: { application/json: { schema: { $ref: '#/components/schemas/Coupon' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/admin/coupons/{couponId}:
    patch:
      tags: [Admin promotions]
      summary: 编辑优惠券草稿字段
      description: 需要 `store.promotions.manage`；发布后的规则版本不可原地修改。
      security: [{ StoreCode: [], AdminBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/CouponId'
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/CouponDraftUpdate' } } }
      responses:
        '200':
          description: 已更新
          content: { application/json: { schema: { $ref: '#/components/schemas/Coupon' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/admin/coupons/{couponId}/status:
    post:
      tags: [Admin promotions]
      summary: 激活、暂停或不可逆结束优惠券
      description: 需要 `store.promotions.publish`、二次确认、乐观锁与 Idempotency-Key。
      security: [{ StoreCode: [], AdminBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/CouponId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/CouponStatusRequest' } } }
      responses:
        '200':
          description: 已转换或幂等重放
          content: { application/json: { schema: { $ref: '#/components/schemas/Coupon' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  securitySchemes:
    StoreCode:
      type: apiKey
      in: header
      name: X-Store-Code
      description: 受信商城业务编码；不能由请求体 store_id 替代
    MemberBearer:
      type: http
      scheme: bearer
      bearerFormat: Zalo-Shop-Member
      description: 商城绑定会员访问令牌；商城必须与 X-Store-Code 一致
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: Zalo-Shop-Admin
      description: 完成密码与 TOTP 的商城管理员令牌
  parameters:
    StoreCode:
      name: X-Store-Code
      in: header
      required: true
      schema: { $ref: '#/components/schemas/Code' }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string, minLength: 1, maxLength: 1000 }
...
    AdminLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 当前商城范围 16–128 字符动作键；同键不同规范请求 hash 冲突
      schema:
        { type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$' }
...
    PromotionId:
      name: promotionId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CouponId:
      name: couponId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: 输入或商城 Header 无效
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Unauthorized:
      description: 认证缺失、过期或无效
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Forbidden:
      description: 商城不匹配、未授权或缺少权限；不泄露资源存在性
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    NotFound:
      description: 当前商城内资源不存在
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Conflict:
      description: 版本、状态、库存或幂等键冲突
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
...
  schemas:
    Code:
      type: string
      minLength: 2
      maxLength: 64
      pattern: '^[a-z][a-z0-9-]*$'
    VndAmount:
      type: integer
      format: int64
      minimum: 0
      maximum: 9007199254740991
...
    Locale:
      type: string
      enum: [vi, zh, en]
    CursorPage:
      type: object
      required: [next_cursor]
      properties:
        next_cursor: { type: [string, 'null'], maxLength: 1000 }
...
    LocalizationInput:
      type: object
      additionalProperties: false
      required: [locale, name]
      properties:
        locale: { $ref: '#/components/schemas/Locale' }
        name: { type: string, minLength: 1, maxLength: 240 }
        description: { type: [string, 'null'], minLength: 1, maxLength: 2000, default: null }
...
    PromotionStatus:
      type: string
      enum: [DRAFT, ACTIVE, PAUSED, ENDED]
    PricingBucket:
      type: string
      enum: [ITEM, ORDER, COUPON, SHIPPING]
...
    PromotionTargetLookupPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            required: [id, code, names]
            properties:
              id: { type: string, format: uuid }
              code: { $ref: '#/components/schemas/Code' }
              names:
                type: object
                additionalProperties: false
                required: [vi, zh, en]
                properties:
                  vi: { type: [string, 'null'], maxLength: 240 }
                  zh: { type: [string, 'null'], maxLength: 240 }
                  en: { type: [string, 'null'], maxLength: 240 }
        next_cursor: { type: [string, 'null'], format: uuid }
    PromotionBenefit:
      oneOf:
        - type: object
          additionalProperties: false
          required: [method, value]
          properties:
            method: { type: string, const: FIXED_VND }
            value: { $ref: '#/components/schemas/VndAmount' }
        - type: object
          additionalProperties: false
          required: [method, value]
          properties:
            method: { type: string, const: PERCENTAGE_BPS }
            value: { type: integer, minimum: 1, maximum: 10000 }
            maximum_discount_vnd:
              { oneOf: [{ $ref: '#/components/schemas/VndAmount' }, { type: 'null' }] }
        - type: object
          additionalProperties: false
          required: [method]
          properties:
            method: { type: string, const: FREE_SHIPPING_QUALIFICATION }
    PromotionVersionInput:
      type: object
      additionalProperties: false
      required:
        [expected_promotion_version, bucket, benefit, starts_at, priority, localizations, targets]
      properties:
        expected_promotion_version: { type: integer, minimum: 1, maximum: 2147483647 }
        bucket: { $ref: '#/components/schemas/PricingBucket' }
        benefit: { $ref: '#/components/schemas/PromotionBenefit' }
        starts_at: { type: string, format: date-time }
        ends_at: { type: [string, 'null'], format: date-time, default: null }
        minimum_spend_vnd: { oneOf: [{ $ref: '#/components/schemas/VndAmount' }, { type: 'null' }] }
        minimum_quantity: { type: [integer, 'null'], minimum: 1, maximum: 99 }
        priority: { type: integer, minimum: 0, maximum: 1000000 }
        stackable_with:
          {
            type: array,
            maxItems: 3,
            uniqueItems: true,
            items: { $ref: '#/components/schemas/PricingBucket' },
          }
        localizations:
          {
            type: array,
            minItems: 1,
            maxItems: 3,
            items: { $ref: '#/components/schemas/LocalizationInput' },
          }
        targets:
          {
            type: array,
            minItems: 1,
            maxItems: 500,
            items: { $ref: '#/components/schemas/PromotionTarget' },
          }
    PromotionVersion:
      type: object
      additionalProperties: false
      required:
        [
          id,
          version_number,
          status,
          bucket,
          benefit,
          starts_at,
          ends_at,
          minimum_spend_vnd,
          minimum_quantity,
          priority,
          stackable_with,
          localizations,
          targets,
          created_at,
          published_at,
        ]
      properties:
        id: { type: string, format: uuid }
        version_number: { type: integer, minimum: 1 }
        status: { type: string, enum: [DRAFT, PUBLISHED] }
        bucket: { $ref: '#/components/schemas/PricingBucket' }
        benefit: { $ref: '#/components/schemas/PromotionBenefit' }
        starts_at: { type: string, format: date-time }
        ends_at: { type: [string, 'null'], format: date-time }
        minimum_spend_vnd: { oneOf: [{ $ref: '#/components/schemas/VndAmount' }, { type: 'null' }] }
        minimum_quantity: { type: [integer, 'null'], minimum: 1, maximum: 99 }
        priority: { type: integer, minimum: 0, maximum: 1000000 }
        stackable_with:
          {
            type: array,
            maxItems: 3,
            uniqueItems: true,
            items: { $ref: '#/components/schemas/PricingBucket' },
          }
        localizations:
          {
            type: array,
            minItems: 1,
            maxItems: 3,
            items: { $ref: '#/components/schemas/LocalizationInput' },
          }
        targets:
          {
            type: array,
            minItems: 1,
            maxItems: 500,
            items: { $ref: '#/components/schemas/PromotionTarget' },
          }
        created_at: { type: string, format: date-time }
        published_at: { type: [string, 'null'], format: date-time }
    Promotion:
      type: object
      additionalProperties: false
      required: [id, code, status, version, active_version, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        code: { $ref: '#/components/schemas/Code' }
        status: { $ref: '#/components/schemas/PromotionStatus' }
        version: { type: integer, minimum: 1 }
        active_version:
          { oneOf: [{ $ref: '#/components/schemas/PromotionVersion' }, { type: 'null' }] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PromotionPage:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/CursorPage'
        - type: object
          required: [items]
          properties: { items: { type: array, items: { $ref: '#/components/schemas/Promotion' } } }
    PromotionVersionPage:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/CursorPage'
        - type: object
          required: [items]
          properties:
            { items: { type: array, items: { $ref: '#/components/schemas/PromotionVersion' } } }
    PublishPromotionRequest:
      type: object
      additionalProperties: false
      required: [version_id, expected_promotion_version, confirmation_code]
      properties:
        version_id: { type: string, format: uuid }
        expected_promotion_version: { type: integer, minimum: 1, maximum: 2147483647 }
        confirmation_code: { type: string, const: PUBLISH }
    PausePromotionRequest:
      type: object
      additionalProperties: false
      required: [expected_promotion_version, confirmation_code]
      properties:
        expected_promotion_version: { type: integer, minimum: 1, maximum: 2147483647 }
        confirmation_code: { type: string, const: PAUSE }
    EndPromotionRequest:
      type: object
      additionalProperties: false
      required: [expected_promotion_version, confirmation_code]
      properties:
        expected_promotion_version: { type: integer, minimum: 1, maximum: 2147483647 }
        confirmation_code: { type: string, const: END }
    CouponInput:
      type: object
      additionalProperties: false
      required: [code, promotion_version_id]
      properties:
        code: { $ref: '#/components/schemas/Code' }
        promotion_version_id: { type: string, format: uuid }
        total_claim_limit: { type: [integer, 'null'], minimum: 1, maximum: 2147483647 }
        per_member_claim_limit: { type: integer, const: 1, default: 1 }
        new_customer_only:
          type: boolean
          default: false
          description: M3 以零订单基线判断；M4 必须使用同商城真实已完成订单事实
    CouponDraftUpdate:
      type: object
      additionalProperties: false
      required: [expected_version]
      minProperties: 2
      properties:
        expected_version: { type: integer, minimum: 1, maximum: 2147483647 }
        promotion_version_id: { type: string, format: uuid }
        total_claim_limit: { type: [integer, 'null'], minimum: 1, maximum: 2147483647 }
        per_member_claim_limit: { type: integer, const: 1 }
        new_customer_only: { type: boolean }
    CouponStatusRequest:
      oneOf:
        - type: object
          additionalProperties: false
          required: [expected_version, status, confirmation_code]
          properties:
            expected_version: { type: integer, minimum: 1, maximum: 2147483647 }
            status: { type: string, const: ACTIVE }
            confirmation_code: { type: string, const: ACTIVATE }
        - type: object
          additionalProperties: false
          required: [expected_version, status, confirmation_code]
          properties:
            expected_version: { type: integer, minimum: 1, maximum: 2147483647 }
            status: { type: string, const: PAUSED }
            confirmation_code: { type: string, const: PAUSE }
        - type: object
          additionalProperties: false
          required: [expected_version, status, confirmation_code]
          properties:
            expected_version: { type: integer, minimum: 1, maximum: 2147483647 }
            status: { type: string, const: ENDED }
            confirmation_code: { type: string, const: END }
    Coupon:
      type: object
      additionalProperties: false
      required:
        [
          id,
          code,
          promotion_version_id,
          total_claim_limit,
          per_member_claim_limit,
          new_customer_only,
          status,
          claimed_count,
          version,
          created_at,
          updated_at,
        ]
      properties:
        id: { type: string, format: uuid }
        code: { $ref: '#/components/schemas/Code' }
        promotion_version_id: { type: string, format: uuid }
        total_claim_limit: { type: [integer, 'null'], minimum: 1 }
        per_member_claim_limit: { type: integer, const: 1 }
        new_customer_only: { type: boolean }
        status: { type: string, enum: [DRAFT, ACTIVE, PAUSED, ENDED] }
        claimed_count: { type: integer, minimum: 0 }
        version: { type: integer, minimum: 1 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CouponPage:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/CursorPage'
        - type: object
          required: [items]
          properties: { items: { type: array, items: { $ref: '#/components/schemas/Coupon' } } }